services/requirements.txt
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
psycopg2==2.9.10
python-dotenv==1.1.0
requests==2.32.3
soupsieve==2.6
typing_extensions==4.13.0
urllib3==2.3.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
asyncpg>=0.28.0
psycopg2-binary 
python-dotenv>=1.0.0
pydantic-settings>=2.0.0